+++ app/pages/index.tsx
@@ -1,0 +1,10 @@
+export default function Home() {
+    return (
+      <div className="min-h-screen bg-black flex items-center justify-center">
+        <button className="bg-white text-black px-6 py-3 rounded-lg font-semibold">
+          Login
+        </button>
+      </div>
+    )
+  }
+
